requirements.txt
Pillow
mongoengine
python-dotenv
kafka-python
mongoengine
pyfiglet
requests
fastapi
python-multipart
uvicorn